[FIX] 출석부 컴포넌트 분리
